check almost ready

+++ PokerBet/Domain/GameBet.cs
@@ -12,12 +12,12 @@ namespace Domain
         public int Id { get; set; }
         public long ContractNumber { get; set; }
 
-        public uint Event { get; set; }
+        public long Event { get; set; }
         public DateTime StartDate { get; set; }
         public int TableNumber { get; set; }
-        public short TableCode { get; set; }
+        public int TableCode { get; set; }
         public int Winner { get; set; }
-        public float Index { get; set; }
+        public double Index { get; set; }
         public int Sum { get; set; }
     }
 }
